update(test): sep test python file by funcs and add Make rules
